voyage-ingest/requirements.txt
google-api-python-client==2.146.0
google-auth==2.35.0
boto3==1.34.140
Pillow>=11.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10

